Update product.c
Change Value to Integer

+++ product.c
@@ -1,4 +1,4 @@
-double product(double num1, double num2)
-{
+// Fix Function Definition
+int product(int num1, int num2) {
   return num1*num2;
 }
